add option for previous sDWM to work with meandering data coming from my meandering process.
Result not Good:

sDWM working with Meandering DATA from my process:
The farm production is: 15344.66 kW, where each turbine is: [4963.7  1214.48 1348.45 1495.48 1559.52 1567.99 1591.17 1603.87]

steady sDWM in progress:
The farm production is: 10047.14 kW, where each turbine is: [4963.7  1214.48  629.26  643.47  666.29  642.52  644.03  643.39]

Observations:
The result degenerate  after the second Turbines...

New purpose:
Understand why and fix my program.
Don't forget that that this previous sDWM seems to be a good approx in term of POWER!
So it's a good thing to have it like a reference.